pyproject.toml
[project]
name = "cs"
version = "0.1.0"
description = "AI Chat Application with Memory - Modern conversation system with persistent history"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "google-genai>=1.20.0",
    "google-generativeai>=0.8.5",
    "requests>=2.32.4",
]

[project.scripts]
cs-chat = "main_new:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pyright]
typeCheckingMode = "off"
include = ["src"]
exclude = ["tests", "docs"]
